pyproject.toml
[project]
name = "open-auric"
version = "0.1.0"
description = "OpenAuric: A local-first, recursive AI agent."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer",
    "textual",
    "fastapi",
    "uvicorn",
    "apscheduler", 
    "litellm",
    "sqlmodel",
    "pydantic-settings",
    "mcp",
    "chromadb",
    "rich",
    "json5",
    "watchdog",
    "aiosqlite",
    "python-telegram-bot",
    "python-telegram-bot",
    "discord.py",
    "psutil",
]

[project.scripts]
auric = "auric.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/auric"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
